I-0014: UPROPERTY TMap 키/값을 uint16 → int32 로 수정 (UE5 reflection 호환)
UnrealHeaderTool 의 UPROPERTY TMap reflection 은 unsigned 정수 중 uint8 만
표준 지원한다. uint16 키/값은 컴파일 단계에서 UHT 에러 또는 reflection
누락 위험. int32 로 일괄 변환하여 안전성 확보.

영향 받은 필드:
- FHktTerrainBakedChunk::SpawnTemplateAttribution
  : TMap<uint16, uint16> → TMap<int32, int32>
- UHktTerrainBakedAsset::SpawnTemplateCatalog
  : TMap<uint16, FGameplayTag> → TMap<int32, FGameplayTag>
- FHktTerrainBakedChunk::PackLocalCoord/UnpackLocalCoord
  : 반환/인자 타입 int32 화 (5+5+5=15bit, 양수 보장)
- UHktTerrainBakedAsset::FindVoxelAttribution 반환 타입
- FHktTerrainProvider::GetChunkVoxelAttribution 의 TPair / 카탈로그 lookup 타입
- BakeRegion 의 TagToTemplateId / catalog accumulator / orphan scan / log format

의미상 templateId 범위 (1..65535) 는 동일. NextTemplateId wraparound 가드도
NextTemplateId > MAX_uint16 로 정정 (이전엔 uint16 wraparound 0 체크).

자동화 테스트는 비-UPROPERTY 영역이라 변경 없음 (uint32 SlotHash31 잔류).

+++ HktGameplay/Source/HktTerrain/Private/HktTerrainBakedAsset.cpp
@@ -176,7 +176,7 @@ void UHktTerrainBakedAsset::GetSpawnersForChunk(const FIntVector& Coord,
 	}
 }
 
-const TMap<uint16, uint16>* UHktTerrainBakedAsset::FindVoxelAttribution(const FIntVector& Coord) const
+const TMap<int32, int32>* UHktTerrainBakedAsset::FindVoxelAttribution(const FIntVector& Coord) const
 {
 	const FHktTerrainBakedChunk* Chunk = FindChunk(Coord);
 	if (!Chunk || Chunk->SpawnTemplateAttribution.Num() == 0)
